Rename manifest-override based on pr feedback
