+++ src/tests/end-to-end/common/manifest-instance.ts
